refactor: split the gateway package into cohesive files
